fix(client): 让 main 重新可构建——PR #20 遗留的 5 处分析器 / 类型错误

+++ modules/input/tests/Unit/InputBudgetRegressionTests.cs
@@ -73,7 +73,7 @@ public sealed class InputBudgetRegressionTests
     {
         for (int i = 0; i < count; i++)
         {
-            ingress.TryEnqueue(new RawInputSample(i + 1, 0, 0));
+            ingress.TryEnqueue(new RawInputSample((uint)(i + 1), 0, 0));
         }
     }
